Supermarket scripts and performance enhancement
