+++ Assets/Code/SupermarketLevel/EntranceDoor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bdd28c210ff92444a7cd84d5091eb3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
