+++ Assets/Code/SupermarketLevel/GrabbableScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be8e517d2a61f5c49ad01d65fb878e28
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
