+++ Assets/Code/SupermarketLevel/PlayerPickUpDrop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acea789bc17156145ab7e3744b4f5692
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
