+++ Assets/Code/SupermarketLevel/ShoppingCartBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b5e15708b61ca9a4ead7f07aefc53617
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
