+++ Assets/Code/SupermarketLevel/movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a14c9d6d5863b4499e61fe27aac8ef2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
