+++ Assets/Scenes/Supermarket/Lightmap-3_comp_dir.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: efeae9edf8110bc479b836299f6d88df
+guid: 19b6d35264c85fd4bb0341c96e20e60d
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
