+++ Assets/Scenes/Supermarket/Lightmap-3_comp_light.exr.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e6295f954e863a747906513eb1a7e120
+guid: 3d186f54d747eb14da9edcd7fb0a25c8
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
