+++ Assets/Scenes/Supermarket/OcclusionCullingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7dffa21be77fd948a7ce8c39fc39556
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 36300000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
