+++ Assets/Scenes/Supermarket/ReflectionProbe-0.exr.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c8053b4d7565a664e81e662b4c95d59a
+guid: 7e6632e8f8c9de94eb2ba7844f9087a1
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
