+++ ProjectSettings/TagManager.asset
@@ -12,7 +12,7 @@ TagManager:
   - Water
   - UI
   - ShoppingCart
-  - 
+  - EntranceDoor
   - 
   - 
   -
